rename entities: plural/singular word
